pyproject.toml
[tool.poetry]
name = "cryptobot"
version = "0.1.0"
description = "package to monitor, alert and do trading with cbpro."
authors = ["Selim Onat <selim.onat@mytaxi.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1"
cbpro = "^1.1.4"
pyyaml = "^6.0"
pandas = "^1.3.4"
matplotlib = "^3.4.3"
dash = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
